production/misc/requirements.txt
numpy==1.17.3
pandas==0.25.3
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
six==1.13.0

